Split image reference normalization
